pyproject.toml
[project]
name = "hypothesis-generation-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "outlines",
    "fastapi",
    "uvicorn[standard]",
    "gunicorn",
    "redis",
    "python-multipart",
    "gseapy",
    "faiss-gpu",
    "python-dotenv",
    "PyJWT==2.9.0",
    "pymongo",
    "torch",
    "torchvision",
    "torchaudio",
    "sentence-transformers",
    "llama_index",
    "llama-index-llms-anthropic",
    "pengines",
    "prefect>=2.14.0,<3.0.0",
    "asyncpg",
    "bokeh>=3.1.0",


    "requests",
    "gwaslab",
    "cyvcf2",
    "polars",
    "python-socketio",
    "werkzeug",
    "tzlocal",
    "PyYAML",
    "psutil",
    "pronto",
    "tiledbsoma",
    "cellxgene-census",
    "loguru>=0.7.3",

    "numpy<2",             
    "optuna>=4.3.0",
    "pandas>=2.2.3",         
    "pyarrow>=17.0.0",       
    "pysam>=0.23.1",
    "ruff>=0.11.12",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "watchdog>=6.0.0",
    "boto3>=1.35.0",
    "prefect-dask",
    "dask[distributed]"
]

[project.optional-dependencies]
r-integration = ["rpy2>=3.5.0,<3.6.0"]